# ViagemAPI/ViagemAPI/ViewModel/LinhaViewModel.cs
namespace ViagemAPI.ViewModel
{
    public class LinhaViewModel
    {
        public int Id { get; set; }
        public string Nome { get; set; }
        public int Numero { get; set; }
        public string Origem { get; set; }
        public string Destino { get; set; }
    }
}